+++ common/cards/base/card.ts
@@ -1,6 +1,4 @@
-import {CardComponent} from '../../components'
 import {TypeT} from '../../types/cards'
-import {DefaultDictionary} from '../../types/game-state'
 import {getCardRank} from '../../utils/ranks'
 import {Card, Hermit, isItem} from './types'
 
@@ -12,21 +10,6 @@ export type CanAttachError =
 	| 'UNKNOWN_ERROR'
 
 export type CanAttachResult = Array<CanAttachError>
-
-/** Type that allows multiple functions in a card to share values. */
-export class InstancedValue<T> extends DefaultDictionary<CardComponent, T> {
-	public set(component: CardComponent, value: T) {
-		this.setValue(component.entity, value)
-	}
-
-	public get(component: CardComponent): T {
-		return this.getValue(component.entity)
-	}
-
-	public clear(component: CardComponent) {
-		this.clearValue(component.entity)
-	}
-}
 
 /** Get the foreground image for a card */
 export function getCardImage(card: Card) {
